+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - add extra filtering on `EXTERNAL_FORMS` via keywords: `LAST_HOUR, LAST_DAY, LAST_WEEK, LAST_MONTH, LAST_YEAR`
  - add option to wrap internal `output_alerts` in container
  - add transaction log when a user is deleted
+ - style `form` add internal confirmation data
 
 ### Bugfix
  -  adjust `entryRecord` to load children like `entryList` and now it properly load values of records in edit mode